🔒 Warden: fix integer overflows in distance calculations (#4924)
- Fixed unhandled integer overflow during distance calculation (e.g. `dx * dx` or `a.x - b.x` causing `attempt to subtract with overflow` and `attempt to multiply with overflow`) allowed panics and DoS if extreme coordinates were passed in `curvature.rs`, `bureaucratic_black_hole.rs`, `radioactive.rs`, `orbital_mirrors.rs`, `hum.rs`, `lighting.rs`, and `seismic.rs`.
- Cast coordinate values to `f32` *before* subtraction or multiplication to ensure floating-point math handles large inputs without panicking.

+++ src/layer1/administration/bureaucratic_black_hole.rs
@@ -56,8 +56,8 @@ pub fn bureaucratic_resource_loss_system(
 ) {
     for (bh, bh_pos) in black_holes.iter() {
         for (entity, res_pos, chance) in resources.iter() {
-            let dx = (bh_pos.x - res_pos.x) as f32;
-            let dy = (bh_pos.y - res_pos.y) as f32;
+            let dx = bh_pos.x as f32 - res_pos.x as f32;
+            let dy = bh_pos.y as f32 - res_pos.y as f32;
             let distance = (dx * dx + dy * dy).sqrt();
 
             if distance <= bh.radius {
@@ -78,8 +78,8 @@ pub fn bureaucratic_pop_reassignment_system(
 ) {
     for (bh, bh_pos) in black_holes.iter() {
         for (entity, pop_pos, chance) in pops.iter() {
-            let dx = (bh_pos.x - pop_pos.x) as f32;
-            let dy = (bh_pos.y - pop_pos.y) as f32;
+            let dx = bh_pos.x as f32 - pop_pos.x as f32;
+            let dy = bh_pos.y as f32 - pop_pos.y as f32;
             let distance = (dx * dx + dy * dy).sqrt();
 
             if distance <= bh.radius {
